Added course deletion and instructor validation

+++ routes/edit-course.js
@@ -2,13 +2,17 @@ const express = require('express');
 const router =  express.Router();
 const db = require('../db.js');
 
-router.get('/', function(req, res, next) {
-    let courseid = req.query.courseid;
-    if(courseid) {
+router.get('/', async function(req, res, next) {
+    const courseid = req.query.courseid;
+    const instructor = res.locals.user.instructor;
+    const userid = res.locals.user._id;
+
+    if(courseid && instructor && await db.utils.isCourseInstructor(courseid, userid)) {
         db.Course.findById(courseid, (err, course) => {
             if(err) {
                 // Redirect to course page and display error message
                 // TODO: Display error message on course page
+                console.log(err);
                 res.redirect('courses');
             } else {
                 let instructorid = res.locals.user._id;
@@ -24,51 +28,57 @@ router.get('/', function(req, res, next) {
         });
     } else {
         // TODO: error message about no course id being provided
-        res.redirect('courses');
+        res.redirect('error');
     }
 });
 
 // Breaking the RESTful rules a bit here. This should be PUT ideally
-router.post('/', function(req, res, next) {
-    // Verify form data is valid and update database
-    let courseid = req.query.courseid;
-    let courseupdate = {
+router.post('/', async function(req, res, next) {
+    // Verify and sanitize form data here
+    const courseid = req.query.courseid;
+    const instructor = res.locals.user.instructor;
+    const userid = res.locals.user._id;
+    const courseupdate = {
         name: req.body.course_name,
         desc: req.body.course_desc,
         visible: req.body.course_visible ? true : false
     };
-    if(courseid) {
+
+    if(courseid && instructor && await db.utils.isCourseInstructor(courseid, userid)) {
         db.Course.findByIdAndUpdate(courseid, courseupdate, (err, course) => {
             if(err) {
+                console.log(err);
                 res.render('edit-course', {
                     course: course,
                     error: "Error modifying course."
                 });
             } else {
-                res.redirect('courses');
+                res.redirect('error');
             }
         })
     } else {
         // TODO: error message about no course id being provided
-        res.redirect('courses');
+        res.redirect('error');
     }
 });
 
-router.delete('/', function(req, res, next) {
-    let courseid = req.query.courseid;
-    if(courseid) {
+router.delete('/', async function(req, res, next) {
+    const courseid = req.query.courseid;
+    const instructor = res.locals.user.instructor;
+    const userid = res.locals.user._id;
+
+    if(courseid && instructor && await db.utils.isCourseInstructor(courseid, userid)) {
         db.Course.findByIdAndRemove(courseid, (err) => {
             if(err) {
-                res.render('edit-course', {
-                    course: course,
-                    error: "Error deleting course."
-                });
+                console.log(err);
+                res.redirect('error');
             } else {
+                // TODO: This redirect call currently does not work in a delete call
                 res.redirect('courses');
             }
         });
     } else {
-        res.redirect('courses');
+        res.redirect('error');
     }
 });
